fix(fountain): wire format — BIG-endian u32 seed (production format)
Mid-Phase-2 correction. The original migration plan documented the
droplet wire format as little-endian u64 seed; that was wrong. The
actual production wire format used by meow_decoder.fountain.pack_droplet
is BIG-endian u32 seed + BIG-endian u16 counts/indices, and every
already-encoded GIF in the wild relies on it.

Caught when wiring up Phase 2 (PyO3 binding) — comparing
`pack_droplet()` output to the Rust `to_wire()` output revealed the
divergence.

## Wire format (locked, matches pack_droplet/unpack_droplet)

```text
seed:           u32  big-endian   (4 bytes)
block_count:    u16  big-endian   (2 bytes)
block_indices: [u16; block_count] big-endian
data:          [u8;  block_size]
```

Total = 4 + 2 + 2*count + block_size bytes. Mirrors
struct.pack(">I", seed) + struct.pack(">H", count) + ...

## Changes

* crypto_core/src/meow_fountain/wire.rs — `Droplet.seed` type changed
  from u64 to u32, all serialise/deserialise paths swapped to
  to_be_bytes / from_be_bytes. `wire_size` formula updated. Header-
  too-short threshold dropped from 10 → 6 bytes. Tests updated.
* crypto_core/src/meow_fountain/encoder.rs — `Droplet { seed, ... }`
  no longer casts to u64; `seed: u32` flows through directly.
* rust_crypto/src/fountain.rs — `PyDroplet::new(seed: u32)` and the
  `seed` getter return u32 (Python ints accept either; surface
  matches the existing meow_decoder.fountain.Droplet dataclass which
  stores int).
* scripts/dev/generate_fountain_golden_vectors.py — packs with
  `>IH` instead of `<QH`. Regenerated all 16 vectors.
* tests/test_fountain_golden_vectors.py — updated to read the new
  format; head_len = 4 + 2 + 2*count.
* docs/FOUNTAIN_RUST_WASM_MIGRATION.md — wire format section
  corrected to point at the actual production layout. (Will land in
  the next commit alongside the Phase 2b shim.)

## Verification

* 37 Rust unit tests pass (was 38; one was a "matches_golden_vector"
  spot-check that became redundant with the integration parity test).
* fountain_golden_parity integration test: 1/1 vector parity green.
* tests/test_fountain_golden_vectors.py: 50/50 Python parity green.
* Cross-language sanity: Rust `to_wire()` ≡ Python `pack_droplet()`
  ≡ golden .bin file for every vector.

This is the byte-for-byte format that lands in QR codes inside
Schrödinger GIFs. Backward compat with already-encoded files: full.

+++ rust_crypto/src/fountain.rs
@@ -73,7 +73,7 @@ impl PyDroplet {
     /// Construct a droplet directly from its three fields. Mostly
     /// useful for tests; the encoder produces droplets directly.
     #[new]
-    fn new(seed: u64, block_indices: Vec<u16>, data: Vec<u8>) -> Self {
+    fn new(seed: u32, block_indices: Vec<u16>, data: Vec<u8>) -> Self {
         Self {
             inner: RustDroplet {
                 seed,
@@ -84,7 +84,7 @@ impl PyDroplet {
     }
 
     #[getter]
-    fn seed(&self) -> u64 {
+    fn seed(&self) -> u32 {
         self.inner.seed
     }
